Move download_file to http_client rather then common
